# backend/swagger.yaml
openapi: 3.1.0
info:
  title: Mixillo API
  version: 1.0.0
  description: |
    Complete API documentation for Mixillo - TikTok-style social commerce platform.
    
    ## Features
    - 🔐 JWT Authentication with refresh tokens
    - 📹 TikTok-style video content & social feeds
    - 🛍️ E-commerce with cart & order management
    - 💰 Digital wallet & coin system
    - 🎁 Virtual gifting & monetization
    - 📺 Live streaming with Agora SDK
    - 👥 Social features (follow, like, comment, share)
    - 📊 Analytics & trending algorithms
    - 🛡️ AI-powered content moderation
    - 📱 Multi-language & multi-currency support
    
  contact:
    name: Mixillo Support
    email: support@mixillo.com
  license:
    name: Proprietary
servers:
  - url: https://mixillo-backend-52242135857.europe-west1.run.app
    description: Production Server (Google Cloud Run)
  - url: http://localhost:5000
    description: Local Development Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authenticated requests
  schemas:
...
    Content:
      type: object
      properties:
        _id:
          type: string
        videoUrl:
          type: string
          example: "https://res.cloudinary.com/mixillo/video/upload/videos/vid123.mp4"
        thumbnailUrl:
          type: string
        caption:
          type: string
          example: "Amazing sunset view! 🌅"
        creator:
          $ref: '#/components/schemas/User'
        likes:
          type: integer
          example: 2345
        comments:
          type: integer
          example: 234
        shares:
          type: integer
          example: 56
        views:
          type: integer
          example: 12345
        duration:
          type: integer
          description: Video duration in seconds
          example: 45
        hashtags:
          type: array
          items:
            type: string
          example: ["#sunset", "#nature", "#amazing"]
        soundId:
          type: string
        isActive:
          type: boolean
          example: true
        moderationStatus:
          type: string
          enum: [pending, approved, rejected]
          example: "approved"
        createdAt:
          type: string
          format: date-time
    
    # ==================== E-commerce Schemas ====================
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
            pagination:
              type: object
              properties:
                total:
                  type: integer
                  example: 1000
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 20
                pages:
                  type: integer
                  example: 50
paths:
...
  /api/content:
    get:
      tags:
        - Content
      summary: Get all content (with filters)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: creatorId
          in: query
          schema:
            type: string
        - name: hashtag
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Content list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    
    post:
      tags:
        - Content
      summary: Create new content (video post)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - video
              properties:
                video:
                  type: string
                  format: binary
                caption:
                  type: string
                hashtags:
                  type: array
                  items:
                    type: string
                soundId:
                  type: string
      responses:
        '201':
          description: Content created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      content:
                        $ref: '#/components/schemas/Content'
...
  /api/products:
    get:
      tags:
        - Products
      summary: Get all products
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: storeId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [-createdAt, price, -price, -rating]
            default: "-createdAt"
      responses:
        '200':
          description: Products list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    
    post:
      tags:
        - Products
      summary: Create new product (seller only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - storeId
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                images:
                  type: array
                  items:
                    type: string
                category:
                  type: string
                storeId:
                  type: string
                stock:
                  type: integer
                variants:
                  type: array
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      product:
                        $ref: '#/components/schemas/Product'
...
  /api/orders:
    get:
      tags:
        - Orders
      summary: Get user orders
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, shipped, delivered, cancelled, refunded]
      responses:
        '200':
          description: Orders list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
    
    post:
      tags:
        - Orders
      summary: Create order from cart
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - shippingAddress
                - paymentMethod
              properties:
                shippingAddress:
                  type: object
                  properties:
                    fullName:
                      type: string
                    phone:
                      type: string
                    street:
                      type: string
                    city:
                      type: string
                    state:
                      type: string
                    zipCode:
                      type: string
                    country:
                      type: string
                paymentMethod:
                  type: string
                  enum: [card, wallet, cash_on_delivery]
                couponCode:
                  type: string
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      order:
                        $ref: '#/components/schemas/Order'
...
  /api/cart:
    get:
      tags:
        - Cart
      summary: Get user cart
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Cart details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      cart:
                        type: object
                        properties:
                          _id:
                            type: string
                          userId:
                            type: string
                          items:
                            type: array
                            items:
                              type: object
                              properties:
                                productId:
                                  type: string
                                name:
                                  type: string
                                price:
                                  type: number
                                quantity:
                                  type: integer
                                image:
                                  type: string
                          subtotal:
                            type: number
                            example: 199.98
...
  /api/admin/users:
    get:
      tags:
        - Admin
      summary: Get all users (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: role
          in: query
          schema:
            type: string
            enum: [user, seller, admin]
        - name: verified
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '403':
          description: Admin access required
...
  /api/admin/analytics:
    get:
      tags:
        - Admin
      summary: Get system analytics (admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: System analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      totalUsers:
                        type: integer
                        example: 10523
                      totalRevenue:
                        type: number
                        example: 245678.50
                      totalOrders:
                        type: integer
                        example: 3456
                      totalContent:
                        type: integer
                        example: 45678
                      activeStreams:
                        type: integer
                        example: 12
        '403':
          description: Admin access required
  
  # ==================== SYSTEM ====================
...